web_qaa_api/requirements.txt
fastapi
uvicorn
transformers
torch
readability-lxml
beautifulsoup4
requests
lxml_html_clean